Replace workflows module as a subpackage
Better code organization and entrypoint handling
