+++ src/aiida_abacus/workflows/__init__.py
@@ -1,0 +1,8 @@
+"""
+Module contains workchains for abacus
+"""
+
+from .base import AbacusBaseWorkChain
+from .relax import AbacusRelaxWorkchain
+
+__all__ = ["AbacusBaseWorkChain", "AbacusRelaxWorkchain"]
